rename: 여행일정 생성 관련 파일 이동
