+++ src/components/createTrip/hashtag/HashtagCreator.tsx
