+++ src/components/createTrip/hashtag/HashtagItem.tsx
