+++ src/components/createTrip/scheduleAccordion/ScheduleAccordion.tsx
