+++ src/components/createTrip/scheduleAccordion/ScheduleAccordionItem.tsx
